# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserPublic {
  id: number;
  username: string;
  email: string;
}

export interface AuthResponse {
  user: UserPublic;
  token: string;
}

export interface Conversation {
  id: number;
  created_at: string;
  other_user_id: number;
  other_username: string;
  other_email: string;
}

export interface ConversationInput {
  recipientId: number;
}

export interface ConversationCreated {
  conversationId: number;
  existing: boolean;
}

export type MessageInputMessageType = typeof MessageInputMessageType[keyof typeof MessageInputMessageType];


export const MessageInputMessageType = {
  text: 'text',
  image: 'image',
  voice: 'voice',
} as const;

export interface MessageInput {
  conversationId: number;
  content: string;
  message_type?: MessageInputMessageType;
  /** @nullable */
  media_data?: string | null;
}

export interface DeleteMessageInput {
  id: number;
}

export interface Message {
  id: number;
  conversation_id: number;
  sender_id: number;
  /** @nullable */
  sender_username?: string | null;
  content: string;
  /** @nullable */
  message_type?: string | null;
  /** @nullable */
  media_data?: string | null;
  created_at: string;
  /** @nullable */
  deleted_for_all?: boolean | null;
}

export type FindUserParams = {
email: string;
};

export type FetchMessagesParams = {
conversationId: number;
};