# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Clone API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: conversations
    description: Conversation operations
  - name: messages
    description: Message operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: User logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials

  /auth/find:
    get:
      operationId: findUser
      tags: [auth]
      summary: Find a user by email
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "404":
          description: User not found

  /conversations:
    get:
      operationId: getConversations
      tags: [conversations]
      summary: Get all conversations for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a new conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationCreated"

  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /messages/list:
    get:
      operationId: fetchMessages
      tags: [messages]
      summary: Get messages for a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /messages/delete-all:
    post:
      operationId: deleteMessageForAll
      tags: [messages]
      summary: Delete a message for all participants
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteMessageInput"
      responses:
        "200":
          description: Message deleted

  /messages/delete-me:
    post:
      operationId: deleteMessageForMe
      tags: [messages]
      summary: Delete a message for the current user only
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteMessageInput"
      responses:
        "200":
          description: Message hidden

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserPublic"
        token:
          type: string

    UserPublic:
      type: object
      required: [id, username, email]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string

    Conversation:
      type: object
      required: [id, created_at, other_user_id, other_username, other_email]
      properties:
        id:
          type: integer
        created_at:
          type: string
        other_user_id:
          type: integer
        other_username:
          type: string
        other_email:
          type: string

    ConversationInput:
      type: object
      required: [recipientId]
      properties:
        recipientId:
          type: integer

    ConversationCreated:
      type: object
      required: [conversationId, existing]
      properties:
        conversationId:
          type: integer
        existing:
          type: boolean

    MessageInput:
      type: object
      required: [conversationId, content]
      properties:
        conversationId:
          type: integer
        content:
          type: string
        message_type:
          type: string
          enum: [text, image, voice]
        media_data:
          type: ["string", "null"]

    DeleteMessageInput:
      type: object
      required: [id]
      properties:
        id:
          type: integer

    Message:
      type: object
      required: [id, conversation_id, sender_id, content, created_at]
      properties:
        id:
          type: integer
        conversation_id:
          type: integer
        sender_id:
          type: integer
        sender_username:
          type: ["string", "null"]
        content:
          type: string
        message_type:
          type: ["string", "null"]
        media_data:
          type: ["string", "null"]
        created_at:
          type: string
        deleted_for_all:
          type: ["boolean", "null"]