# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPublic } from './userPublic';

export interface AuthResponse {
  user: UserPublic;
  token: string;
}