# lib/api-zod/src/generated/types/conversationCreated.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConversationCreated {
  conversationId: number;
  existing: boolean;
}