# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  conversation_id: number;
  sender_id: number;
  /** @nullable */
  sender_username?: string | null;
  content: string;
  /** @nullable */
  message_type?: string | null;
  /** @nullable */
  media_data?: string | null;
  created_at: string;
  /** @nullable */
  deleted_for_all?: boolean | null;
}