# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageInputMessageType } from './messageInputMessageType';

export interface MessageInput {
  conversationId: number;
  content: string;
  message_type?: MessageInputMessageType;
  /** @nullable */
  media_data?: string | null;
}