# lib/api-zod/src/generated/types/userPublic.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserPublic {
  id: number;
  username: string;
  email: string;
}